Test edits

+++ robottelo/hosts.py
@@ -137,6 +137,7 @@ class ContentHost(Host, ContentHostMixins):
             # key file based authentication
             kwargs.update({'key_filename': auth})
         self._satellite = kwargs.get('satellite')
+        self.blank = kwargs.get('blank', False)
         super().__init__(hostname=hostname, **kwargs)
 
     @property
@@ -195,13 +196,15 @@ class ContentHost(Host, ContentHostMixins):
         return Version(version=version_string)
 
     def setup(self):
-        self.remove_katello_ca()
-        self.execute('subscription-manager clean')
+        if not self.blank:
+            self.remove_katello_ca()
+            self.execute('subscription-manager clean')
 
     def teardown(self):
-        if self.nailgun_host:
-            self.nailgun_host.delete()
-        self.unregister()
+        if not self.blank:
+            if self.nailgun_host:
+                self.nailgun_host.delete()
+            self.unregister()
         # Strip most unnecessary attributes from our instance for checkin
         keep_keys = set(self.to_dict()) | {'release', '_prov_inst'}
         self.__dict__ = {k: v for k, v in self.__dict__.items() if k in keep_keys}
